moving helper tests to helpers, adding stream tag conversion support
